chore: remove unused configs
