+++ nvim/.config/nvim/lua/plugins/autocompletions.lua
