+++ nvim/.config/nvim/lua/plugins/fterm.lua
