+++ nvim/.config/nvim/lua/plugins/lspconfig.lua
