+++ nvim/.config/nvim/lua/plugins/lualine.lua
