+++ nvim/.config/nvim/lua/plugins/mason.lua
