+++ nvim/.config/nvim/lua/plugins/neotree.lua
