+++ nvim/.config/nvim/lua/plugins/nvim-tmux-navigator.lua
